Fix moundId type.

+++ src/world/Action/MountAction.h
@@ -8,7 +8,7 @@ namespace Sapphire::World::Action
   class MountAction : public Action
   {
   public:
-    MountAction( Entity::CharaPtr source, uint32_t mountId, uint16_t sequence, Data::ActionPtr actionData, FrameworkPtr fw );
+    MountAction( Entity::CharaPtr source, uint16_t mountId, uint16_t sequence, Data::ActionPtr actionData, FrameworkPtr fw );
     virtual ~MountAction() = default;
 
     bool preCheck() override;
@@ -18,7 +18,7 @@ namespace Sapphire::World::Action
     void execute() override;
 
   private:
-    uint32_t m_mountId;
+    uint16_t m_mountId;
   };
 }
